refactor(subscription): replace SubscriptionInheritanceService with direct function calls for improved clarity and maintainability in billing and feature usage logic
